Changed messages from sprite to text
Adapted existing message function to display text rather than sprites.
Sprites would be ideal, but we do not have sprites for any of the
messages we need to print. It can easily be changed back to sprites in
the future.

Added a sample message after the initial rolloff to say who goes first.