Goal Scoring and changes to pass card
Enabled scoring by passing in to goal. This starts a rolloff. Success
means a goal is scored and the next turn starts. Failure means the
defense gets the ball and does a "goal kick" their next turn.

Added the ai for doing goal kicks. Did not add goal kicks for human
player.

Changed pass so that when the human player is rolling to pass the ball,
the player must click the screen to roll the dice. Previously, the dice
rolled automatically. Did the same for the rolloff when trying to score
with a pass.